+++ Assets/LDtkUnity/Editor/Utility/Tree/LDtkTreeView.cs
@@ -1,7 +1,14 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using UnityEditor.IMGUI.Controls;
 using UnityEngine;
+
+#if UNITY_6000_2_OR_NEWER
+    using TreeView = UnityEditor.IMGUI.Controls.TreeView<int>;
+    using TreeViewItem = UnityEditor.IMGUI.Controls.TreeViewItem<int>;
+    using TreeViewState = UnityEditor.IMGUI.Controls.TreeViewState<int>;
+#else
+using UnityEditor.IMGUI.Controls;
+#endif
 
 namespace LDtkUnity.Editor
 {