+++ Assets/LDtkUnity/Editor/Utility/Tree/LDtkTreeViewProject.cs
@@ -1,4 +1,9 @@
-﻿using UnityEditor.IMGUI.Controls;
+﻿#if UNITY_6000_2_OR_NEWER
+using TreeViewItem = UnityEditor.IMGUI.Controls.TreeViewItem<int>;
+using TreeViewState = UnityEditor.IMGUI.Controls.TreeViewState<int>;
+#else
+using UnityEditor.IMGUI.Controls;
+#endif
 
 namespace LDtkUnity.Editor
 {